feat: BUZZER touches in MID BIP mode + led plays each 5000 seconds when MessageResultt is NULL (it means it's waiting

+++ src/Modules/BUTTONS/BUTTONElectricityModule.h
@@ -5,7 +5,7 @@
 // #include "Modules/LIGHT_SENSORS/LIGHT_SENSOR_TASK.h"
 
 
-uint8_t CHECK_BUTTON_PHYSICAL_STATS(int PIN_ID);
+uint8_t CHECK_BUTTON_PHYSICAL_STATS(uint8_t PIN_ID);
 void ON_BUTTON_PRESSED(void); 
 
 #endif
