+++ src/Modules/GLOBAL/ISRFunctionsCallback.h
@@ -6,7 +6,7 @@
 #include <avr/io.h>
 #include <stdint.h> 
 
-extern volatile int TimesPressedVAR; 
+extern volatile uint8_t TimesPressedVAR; 
 extern LED_PINS LEDS[]; 
 extern void (*REACTORS[])(volatile uint8_t *PORT_ID, uint8_t PIN_ID);
